databaseInit to a separate file
